Okay for today
While I'm a programmer, I also want to make the game, and right now
the most practical thing is to render internally at a relatively low
resolution, and abused downscaling as "optimization", as long as we
hit above 30fps on my machine on debug, we are good to go.

I do wish to optimize this game, however I do not have time in this
moment, I want to focus on other stuff first. Like the game itself.

Only postprocessing seems to murder framerate though, the rest of the
stuff is fine.