graphics_assets
NOTE(jerry): there is a large choice to not use the graphics_assets
system as the default path, it's just an option to have managed assets.

This is because I can clone screenbuffers which are not intended to be
graphics assets per say.

I can turn them into handles by reserving texture slots for them,
which I may consider, but right now this allows more freedom.