Weather that gets the point across for snow and rain
(might have to tweak around values to get it to be more visible (or
lower the game resolution...))

Either way okay, that's a start. We might need better tilesets for
more visible weather but it is there. Or I could try to thicken the
line a little.

I'm okay with this progress for now...

Tomorrow, area transitions! Then an introductory dialogue system.

Basically my plan is to get a draft of all the game systems so I can
kind of see where the game will be going. Then we'll whiteboard draft
the intro, and then polish up systems from there.

Combat will likely be the last thing that is programmed for the draft
stage as it's the most involved part.