Save menu done, now need a load menu
Grr I need more stuff that works with length strings natively

This is literally going to be like 90% of the work today, since the
rest is way easier editor work.

This is basically UI programming and requires me to handle a lot of
state, and NO I'm not writing an IMGUI for something like this. It's
way too much work for a relatively simple game...

The game itself is likely going to use a fully custom UI while the
editor needs something a bit more standard.

UI will always kind of suck...