okay level transitions have been serialized
So that's a majority of the editor's groundwork, since these
primitives can be reused for everything else.

Now I'll just get level transitions to work in game and we
can move on to getting some NPCs with dialogue to work...

These will not be serialized yet.

Clean up will probably happen on friday or something