Finally adding the combat menus
It's a Monday?

Anyways once this is done, we'll set up the proposal for scripts,
which should be pretty easy. Unfortunately the editing workflow
is going to suffer heavily, but I guess that's not too big of a deal.

I don't have infinite time to make nice tools...

We'll try to code our first special ability. Which should be weapon
limited but whatever for now.

I NEED MORE SPRITES, AND I NEED TO FIND TIME TO DRAW SOME CRAPPY ONES...