Hacks to get stuff working
It's a small one, and the proper fix would be to use a stack
allocator.

Which I don't use so I'll just do this for now and keep note that
a real stack allocator would be useful, for now we're just taking
advantage of static memory to do this since it's a small fix and not
exactly the worst thing I've ever had to do.