Tile layers
We're choosing a sane default of only considering collision on ground
and object layers, while we currently decide to allow the tile to
decide if it's collidable or not.

We may change this even further, but for now this is a pretty good
default, and is backwards compatible with the old files.