some fixes I guess
okay, I'm stopping for now, but when I'm back I need to do some
inventory changes, (IE equipment screen mostly), once that's good
we'll move on to trying to animate some basic combat things(first
abilities with animations) and working on some initial combat formulas.

Once that's good we'll get a levelling system. I won't be able to
finish this before I move, so I'll move this project to github as
a private repo (don't have time to/feel like set(ting) up a git server)

Ideally I'd have something playable at the end of the month for an Act
1.

Game is not expected to be more than 2:30 hours or something like
that. Don't have time for all that content creation.

I want this game to be more than just an engine demo, I do want this
to be a real game with a semi coherent plot.

The dialogue system is under consideration for revamp and obviously
the storyboard needs some fixing.