Fixed closing deadlock
Make sure we don't try to wait for thread job completion when we're
all done.

I don't want to use a spinlock for job synchronization, and I should
most certainly use a real queue, to avoid having to use
mutexes... It's technically still faster than the non-multithreaded
version and I suspect when I multithread the command system, it'll be
dramatically faster.