Introducing the shop system stuff.
The cutscene stuff is frankly more vague, and I think I need to spend
a little more time writing some story/script stuff.

This is a bit more concrete which is why I'm working on this now since
I know I will actually have results, and clear marks for it.

Will do cutscene stuff later this week whenever I can find time.

I only wrote the "kernel" for shopping, and I need to code the
UI (which using my methodology is probably not going to be the record
for UI creation/prototyping.)

I'll probably work a little on the UI and doing more stuff
later. (I'll have the states for it setup...)

Also I think my art situation is going to be more dire as doing pixel
art is already a little uncomfortable, and while this is an excellent
distraction-free environment.

It is not conducive to drawing pixel art.

I need a drawing tablet.