building more infrastructure.
After this, I need to start animating some stuff a bit more, and more
minor work.

Then move stuff to files very slowly. Level design draft, and have
enough to start some rudimentary combat.

That's going to be a bit difficult to come up with decent AI
for. Basically this is the *combat* week, in terms of fleshing out
fundamental mechanics...

All current work is making sure the combat system I'm planning is
implemented and working.