Doing more data driven things.
Going to add items from files.

While the format is likely to change, that's okay. I might as well
have the structure setup now, since afterwards my life will assuredly
be a lot easier...

Basically I want to mostly worry about content related stuff, and most
of the work that I'm doing that isn't related to the combat system
should be dedicated to making my life easier in terms of content creation.

The level editor unfortunately will require way more time to be
ergonomic but I can intentionally force the game to be way more linear
so that this isn't a real problem...

Anyways bathroom break. I'm doing a little better on the time...