slowly adding more data parsing...
I now have an existential crisis on whether this even matters? I could
just hard code the items and it would probably still be okay, since
the thing that really matters is the game mechanics... But then again
this is an RPG which means the data is substantially larger than most
games so this makes sense.

But is this premature optimization?

.... IDK why I'm thinking this. I was planning on doing this later
already obviously...