slowly migrating changes to allow for combat selection
Only ability selection for now (targetting stuff), it's not full but
enough for the combat skills. Then I'll need to add the inventory
stuff soon. So we'll work on the combat stuff.

Although I'm going to fix the draw order problem since it's not super
difficult to do a localized thing. (We're not going to sort render
commands, I'll just make an intermediary to get the effects since we
have a generous amount of scratch memory and I don't mind sorting this
every frame...)