Cool stuff is working
Just getting abilities to play.

Once we get sword rush to work we'll work on the post battle UI and
stuff like that.

We'll see what's next on the list... But we should be able to draft
out the combat scenario for act 1 pretty soon.

Although I fear the code rot might be getting to everything... That
can be pretty bad for me.

We might just work on a basic particle system or something after the
UI?

Trying to think about how that might work... But again this is
supposed to be around the complexity of an SNES era game so keep that
limitation in mind.

In that logic, the core game engine is nearly complete...