Sortable draw entities working
Cool, now should I have used the render commands sort? Well possibly,
but this is more explicit thing since it's pretty clear that I never
have to sort the vast majority of items.

Even if insertion sort is basically free on already sorted items,
might as well cull it down to absolutely what I'm required to sort,
which is everything else.

Even better is I should cache the level layers to a bitmap, but that
might be a different form of expensive so I should best not.

Anyways I'm running my time limit a bit so let's try to get some
particles in.

Yes it's a bit of some cowboy coding mess, mostly in organization but
not in code per say (just file weirdos.)