NVM, I seem to be good at small mistakes
Happens all the time I guess. Anyways the scaffolding required to
implement a simple save system based on deltas is now ready, we just
need to start serializing more and more state.

Just add it as a TODO, as I'm not actively looking to work on the save
system. Just trying to fill in more holes that I will inevitably
working on later now.

Although most of the recent UI that has been constructed as of late is
definitely going to be kept.