Cool changes should be working
Now I need to update the ability format to allow for a
dependency block field

(requires)

NOTE: This is because the ability sequence format is not game script,
and without implementation of conditionals this is the simplest way to
get a compilable format.

(requires) just sets some dependencies that must be met for any
remaining blocks to execute unless it is another requires.

We also have specific entity targets now which is a lot of fun,
(Because there is no such thing as looping yet, I mean this is
basically just a weird assembler, so I could implement looping but
let's avoid doing that...)