Regularizing some code.
I'm thinking of how to solve my serialization issue, and I'm probably
just going to resign to serializing whatever is possible to automate,

and manually managing the few spots that have to be manual.

This mainly means, anything that requires unpacking is still manually
fixed out, which is fine since only particularly big structs (like the
entity), or structs that have runtime specific data are modified now.

Which I think is a fair compromise.

The entities are a very special case because they maintain a more
specialized structure than most of the other stuff and I'm too lazy to
make an actual array type for everything else...