Position marker beginnings
A quick thing mostly to use for cutscene purposes.

Also I realize how inconsistent my codebase is.

In fact it's so inconsistent you would think there were like 5
different programmers sometimes.

I swear it's just me. It's just I also use this as a sandbox for
programming (even though this is a serious project.) It's maintainable
to me so that's fine...