Working on the credits screen module thing
Today I'm just going to add a lot of UI and revise the UI.

As much as I would like to push the game into 320x240. It was kind of
a mistake that I didn't realize this for some reason. But at least it
allow the engine to theoretically move to actual 32x32 tiles in the
future for a high resolution "update".

The performance on the web build is really the most concerning part,
since on desktop I have pretty high confidence it will run on most
computers that it has to run on.

(Bloom and Mode7 are the most expensive parts of the rendering. The
shaders perhaps as well...)