[needsimpl]Introduced COMBAT_ONLY flag and entity restrictions
So we can lock items to be of use to only certain entities.

Future commits will have to be smaller since my attention span seems
to be bad when I work on this project most of the time, so I need this
in bite-sized chunks so I don't forget to finish it!